Implemented the budgets service

+++ src/main/kotlin/main/business/transactions/dto/EditTransactionDto.kt
@@ -4,7 +4,6 @@ import jakarta.validation.constraints.NotBlank
 import jakarta.validation.constraints.NotNull
 import main.business.transactions.enums.TransactionType
 import main.business.categories.repo.Category
-import java.math.BigDecimal
 import java.time.LocalDateTime
 
 class EditTransactionDto {
@@ -16,7 +15,7 @@ class EditTransactionDto {
         var userId: Long? = null
 
         @NotNull
-        var amount: BigDecimal? = null
+        var amount: Double? = null
 
 
         var description: String? = null
